data/requirements.txt
Pillow==2.8.2
argparse==1.2.1
boto==2.38.0
cloudinary==1.1.1
requests==2.7.0
wsgiref==0.1.2
